backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
drf-spectacular==0.28.0
celery==5.4.0
redis==5.2.1
django-redis==5.4.0
psycopg[binary]==3.2.3
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
gunicorn==23.0.0
python-decouple==3.8
Pillow==11.1.0
